fix/ delete files task

+++ server/controllers/fsController.js
@@ -7,11 +7,17 @@ async function deleteFile (file){
     console.log(file)
     try {
         const fileRoot = path.join(root, file);
-        await fs.promises.unlink(fileRoot)
-        logger.info(`[fsCrontroller]: delete file ${file} successful`)
+        if ( fs.existsSync(fileRoot) ){
+            await fs.promises.unlink(fileRoot)
+            logger.info(`[fsCrontroller]: delete file ${file} successful`)
+
+        }else{
+            logger.warn(`[fsCrontroller]: couldn't found ${file}`)
+        }
     }catch(err){
+        console.log(err)
         logger.error(`[fsCrontroller]: couldn't delete ${file}`)
-        throw new Error()
+        throw new Error(err)
     }
 }
